services/web/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
Flask-Bootstrap 
Flask-FlatPages
Flask-wtf
Flask-Login
Flask-Uploads
Flask-Admin
Flask-Script
Flask-Migrate
Flask-Mail
Flask-RESTful
Flask-HTTPAuth
Flask-Session
Flask-Caching
Flask-Compress
Flask-Gravatar
Flask-Principal
Flask-SeaSurf
Flask-SSLify
Flask-Assets
Flask-PageDown
Flask-Images
pyyaml
